Move CommonMetadata to its own module
